renaming test to rfdiscovery
