+++ modules/rfdiscovery/proto/rfdiscovery.proto
@@ -1,25 +1,23 @@
-/* capmc.proto: describes the CAPMCConfig object
+/* rfdiscovery.proto: describes the rfdiscovery object
  *
- * Author: Ghazanfar Ali, ghazanfar.ali@ttu.edu; J. Lowell Wofford <lowell@lanl.gov>
+ * Author: Ghazanfar Ali <ghazanfar.ali@ttu.edu>, J. Lowell Wofford <lowell@lanl.gov>
  *
  * This software is open source software available under the BSD-3 license.
  * Copyright (c) 2018, Triad National Security, LLC
  * See LICENSE file for details.
  */
 
- syntax = "proto3";
- package proto;
+syntax = "proto3";
+package proto;
 
- message RFDiscoveryConfig { // information of the redfish discovery module
-    map<string, RFAggregator> servers = 1;
-    string polling_interval = 2;
-    string name_url = 3;
-    string server_url = 4;
-    string uuid_url = 5;
-    string ip_url = 6;
+message RFDiscoveryConfig {
+    map<string, RFDiscoveryServer> servers = 1;
+    string ip_url = 2;
+    string agg_url = 3;
+    string polling_interval = 4;
 }
 
-message RFAggregator { // information of the aggregator server
+message RFDiscoveryServer {
     string name = 1;
     string ip    = 2;
     int32 port = 3;
