+++ modules/rfdiscovery2/proto/rfdiscovery2.proto
@@ -1,0 +1,26 @@
+/* capmc.proto: describes the CAPMCConfig object
+ *
+ * Author: Ghazanfar Ali, ghazanfar.ali@ttu.edu; J. Lowell Wofford <lowell@lanl.gov>
+ *
+ * This software is open source software available under the BSD-3 license.
+ * Copyright (c) 2018, Triad National Security, LLC
+ * See LICENSE file for details.
+ */
+
+ syntax = "proto3";
+ package proto;
+
+ message RFDiscoveryConfig { // information of the redfish discovery module
+    map<string, RFAggregator> servers = 1;
+    string polling_interval = 2;
+    string name_url = 3;
+    string server_url = 4;
+    string uuid_url = 5;
+    string ip_url = 6;
+}
+
+message RFAggregator { // information of the aggregator server
+    string name = 1;
+    string ip    = 2;
+    int32 port = 3;
+}
